Track more information when listening to events
Sometimes, as the main thread is not waiting for the worker thread to finish, it may be in the middle of an action (e.g. sending a key press "A"). If it does not get to finish the action (send a key *release* "A"), "A" stays pressed after the script terminates, which causes issues.
With the additional logging, we can say which keys were pressed but not released, and take care of them manually afterwards.